create BaseAddDatabasePartitionsCommand.
